+++ kubemarine/patches/__init__.py
@@ -22,7 +22,7 @@ The whole directory is automatically cleared and reset after new version of Kube
 from typing import List
 
 from kubemarine.core.patch import Patch
-from kubemarine.patches.p3_enable_calico_audit import EnableCalicoAudit
+from kubemarine.patches.p1_enable_calico_audit import EnableCalicoAudit
 
 patches: List[Patch] = [
     EnableCalicoAudit(),
